# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeMarket API",
    "version": "1.3.0",
    "description": "WeMarket QR Menu & Small Business Platform API"
  },
  "servers": [
    {
      "url": "https://wemarket.onrender.com/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "orderCapability": {
        "type": "apiKey",
        "in": "header",
        "name": "x-order-capability"
      }
    }
  },
...
  "paths": {
...
    "/api/weather/current": {
      "get": {
        "tags": [
          "Weather"
        ],
        "summary": "현재 날씨 조회",
        "parameters": [
          {
            "in": "query",
            "name": "stn",
            "schema": {
              "type": "string"
            },
            "description": "기상관측 지점번호 (기본값: 108 서울)"
          }
        ],
        "responses": {
          "200": {
            "description": "현재 날씨 정보"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/weather/enhanced": {
      "get": {
        "tags": [
          "Weather"
        ],
        "summary": "향상된 날씨 컨텍스트 조회 (추천 시스템용)",
        "parameters": [
          {
            "in": "query",
            "name": "stn",
            "schema": {
              "type": "string"
            },
            "description": "기상관측 지점번호 (기본값: 108 서울)"
          }
        ],
        "responses": {
          "200": {
            "description": "향상된 날씨 컨텍스트 (체감온도, 공기질, 계절, 시간대, 음식 가중치 포함)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/weather/coords": {
      "get": {
        "tags": [
          "Weather"
        ],
        "summary": "좌표 기반 날씨 조회",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "위도"
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "경도"
          }
        ],
        "responses": {
          "200": {
            "description": "좌표 기반 날씨 정보"
          }
        }
      }
    }
  }
}